requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# AI and API dependencies
openai==1.3.0
requests==2.31.0

# Image processing dependencies (Android/Termux compatible)
numpy>=1.21.0,<1.25.0  # Compatible range for Android
Pillow>=9.0.0

# Professional AI upscaling - Real-SR (heavy dependencies - see requirements_realsr.txt)
# realesrgan==0.3.0  # Requires PyTorch, torchvision, OpenCV - moved to separate file
# basicsr==1.4.2     # Heavy ML dependencies - moved to separate file

# Optional: Face enhancement (comment out if installation fails on Termux)
# gfpgan==1.3.8

# Development and utilities
python-dotenv==1.0.0